requirements.txt
google-generativeai
Pillow 